+++ libs/maximilian.h
@@ -217,7 +217,7 @@ private:
     maxiLagExp<double> loopRecordLag;
     
 public:
-    int	myDataSize;
+    unsigned long myDataSize;
     short 	myChannels;
     int   	mySampleRate;
     long length;
@@ -345,7 +345,7 @@ public:
     char *getSummary()
     {
         char *summary = new char[250];
-        sprintf(summary, " Format: %d\n Channels: %d\n SampleRate: %d\n ByteRate: %d\n BlockAlign: %d\n BitsPerSample: %d\n DataSize: %d\n", myFormat, myChannels, mySampleRate, myByteRate, myBlockAlign, myBitsPerSample, myDataSize);
+        sprintf(summary, " Format: %d\n Channels: %d\n SampleRate: %d\n ByteRate: %d\n BlockAlign: %d\n BitsPerSample: %d\n DataSize: %ld\n", myFormat, myChannels, mySampleRate, myByteRate, myBlockAlign, myBitsPerSample, myDataSize);
         std::cout << myDataSize;
         return summary;
     }
